refactor: file -> root/file.c
